Updated metrics to start each portion of pipeline only if data to do so is available
Additionally, modified extraction of labels from scores so that provided labels are supplemented with derived labels as long as a threshold is given
and the corresponding predictor name was not given in y_label_paths. Before if any provided labels were given, the calculation of derived labels was skipped
completely.